+++ babelfish-api/src/main/java/de/conterra/babelfish/plugin/v10_11/feature/builder/QueryBuilder.java
@@ -17,7 +17,7 @@ import java.util.Set;
  * defines a builder of feature {@link Query} requests
  *
  * @author ChrissW-R1
- * @version 0.2.0
+ * @version 0.4.0
  * @since 0.2.0
  */
 public class QueryBuilder {
@@ -45,8 +45,8 @@ public class QueryBuilder {
 	 * @throws QueryParseException if an {@link Exception} occurred, while execute the {@link Query}
 	 * @since 0.1.0
 	 */
-	public static <T extends FeatureObject> ObjectValue build(Layer<T> layer, CoordinateReferenceSystem crs, GeometryObject geometry, String whereClause, Set<? extends Integer> featureIds, boolean idsOnly, boolean countOnly)
-			throws QueryParseException {
+	public static <T extends FeatureObject> ObjectValue build(Layer<T> layer, CoordinateReferenceSystem crs, GeometryObject geometry, String whereClause, Set<? extends Long> featureIds, boolean idsOnly, boolean countOnly)
+	throws QueryParseException {
 		ObjectValue result = de.conterra.babelfish.plugin.v10_02.feature.builder.QueryBuilder.build(layer, crs, geometry, whereClause, featureIds, idsOnly, countOnly);
 		
 		result.addContent("hasZ", new BooleanValue((new LayerWrapper<>(layer)).getEnvelope().getDimension() >= 3), "fields", false);
